feat(plugin-tailwind): automatically read tailwind config file
