+++ tests/integration/tailwindcss/fixtures/tailwindcss-v2/modern.config.ts
