+++ tests/integration/tailwindcss/fixtures/tailwindcss-v2/package.json
