+++ tests/integration/tailwindcss/fixtures/tailwindcss-v3-js-config/modern.config.ts
