+++ tests/integration/tailwindcss/fixtures/tailwindcss-v3-js-config/src/App.jsx
@@ -4,7 +4,7 @@ import 'tailwindcss/utilities.css';
 
 const App = () => (
   <div className="h-12 w-48">
-    <p className="text-xl font-medium text-black">hello world</p>
+    <p className="text-xl font-medium text-red-500">hello world</p>
   </div>
 );
