+++ tests/integration/tailwindcss/fixtures/tailwindcss-v3-js-config/tailwind.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  theme: {
+    colors: {
+      'red-500': '#1fb6ff',
+    },
+  },
+};
