+++ tests/integration/tailwindcss/fixtures/tailwindcss-v3-ts-config/package.json
@@ -1,0 +1,14 @@
+{
+  "private": true,
+  "name": "integration-tailwindcss-v3-ts-config",
+  "version": "2.9.0",
+  "dependencies": {
+    "@modern-js/app-tools": "workspace:*",
+    "@modern-js/plugin-tailwindcss": "workspace:*",
+    "@modern-js/runtime": "workspace:*",
+    "postcss": "8.4.27",
+    "tailwindcss": "^3.3.3",
+    "react": "^18",
+    "react-dom": "^18"
+  }
+}
