+++ tests/integration/tailwindcss/fixtures/tailwindcss-v3-ts-config/tailwind.config.ts
@@ -1,0 +1,7 @@
+export default {
+  theme: {
+    colors: {
+      'red-500': '#1fb6ff',
+    },
+  },
+};
