+++ tests/integration/tailwindcss/fixtures/tailwindcss-v3/src/App.jsx
@@ -1,0 +1,11 @@
+import 'tailwindcss/base.css';
+import 'tailwindcss/components.css';
+import 'tailwindcss/utilities.css';
+
+const App = () => (
+  <div className="h-12 w-48">
+    <p className="text-xl font-medium text-red-500">hello world</p>
+  </div>
+);
+
+export default App;
