+++ tests/integration/tailwindcss/tests/tsconfig.json
